// packages/amplify-graphql-api-construct-tests/src/__tests__/graphql-schemas/reference-style-dynamic-group-auth/API.ts
export type CreatePrimaryMutationVariables = {
  input: CreatePrimaryInput,
  condition?: ModelPrimaryConditionInput | null,
};

export type CreatePrimaryMutation = {
  createPrimary?:  {
    __typename: "Primary",
    id: string,
    content?: string | null,
    groups?: Array< string | null > | null,
    relatedMany?:  {
      __typename: "ModelRelatedManyConnection",
      items:  Array< {
        __typename: "RelatedMany",
        id: string,
        content?: string | null,
        groups?: Array< string | null > | null,
        primaryId?: string | null,
        primary?:  {
          __typename: "Primary",
          id: string,
          content?: string | null,
          groups?: Array< string | null > | null,
          createdAt: string,
          updatedAt: string,
        } | null,
        createdAt: string,
        updatedAt: string,
      } | null >,
      nextToken?: string | null,
    } | null,
    relatedOne?:  {
      __typename: "RelatedOne",
      id: string,
      content?: string | null,
      groups?: Array< string | null > | null,
      primaryId?: string | null,
      primary?:  {
        __typename: "Primary",
        id: string,
        content?: string | null,
        groups?: Array< string | null > | null,
        relatedMany?:  {
          __typename: "ModelRelatedManyConnection",
          nextToken?: string | null,
        } | null,
        relatedOne?:  {
          __typename: "RelatedOne",
          id: string,
          content?: string | null,
          groups?: Array< string | null > | null,
          primaryId?: string | null,
          createdAt: string,
          updatedAt: string,
        } | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type UpdatePrimaryMutationVariables = {
  input: UpdatePrimaryInput,
  condition?: ModelPrimaryConditionInput | null,
};

export type UpdatePrimaryMutation = {
  updatePrimary?:  {
    __typename: "Primary",
    id: string,
    content?: string | null,
    groups?: Array< string | null > | null,
    relatedMany?:  {
      __typename: "ModelRelatedManyConnection",
      items:  Array< {
        __typename: "RelatedMany",
        id: string,
        content?: string | null,
        groups?: Array< string | null > | null,
        primaryId?: string | null,
        primary?:  {
          __typename: "Primary",
          id: string,
          content?: string | null,
          groups?: Array< string | null > | null,
          createdAt: string,
          updatedAt: string,
        } | null,
        createdAt: string,
        updatedAt: string,
      } | null >,
      nextToken?: string | null,
    } | null,
    relatedOne?:  {
      __typename: "RelatedOne",
      id: string,
      content?: string | null,
      groups?: Array< string | null > | null,
      primaryId?: string | null,
      primary?:  {
        __typename: "Primary",
        id: string,
        content?: string | null,
        groups?: Array< string | null > | null,
        relatedMany?:  {
          __typename: "ModelRelatedManyConnection",
          nextToken?: string | null,
        } | null,
        relatedOne?:  {
          __typename: "RelatedOne",
          id: string,
          content?: string | null,
          groups?: Array< string | null > | null,
          primaryId?: string | null,
          createdAt: string,
          updatedAt: string,
        } | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type DeletePrimaryMutationVariables = {
  input: DeletePrimaryInput,
  condition?: ModelPrimaryConditionInput | null,
};

export type DeletePrimaryMutation = {
  deletePrimary?:  {
    __typename: "Primary",
    id: string,
    content?: string | null,
    groups?: Array< string | null > | null,
    relatedMany?:  {
      __typename: "ModelRelatedManyConnection",
      items:  Array< {
        __typename: "RelatedMany",
        id: string,
        content?: string | null,
        groups?: Array< string | null > | null,
        primaryId?: string | null,
        primary?:  {
          __typename: "Primary",
          id: string,
          content?: string | null,
          groups?: Array< string | null > | null,
          createdAt: string,
          updatedAt: string,
        } | null,
        createdAt: string,
        updatedAt: string,
      } | null >,
      nextToken?: string | null,
    } | null,
    relatedOne?:  {
      __typename: "RelatedOne",
      id: string,
      content?: string | null,
      groups?: Array< string | null > | null,
      primaryId?: string | null,
      primary?:  {
        __typename: "Primary",
        id: string,
        content?: string | null,
        groups?: Array< string | null > | null,
        relatedMany?:  {
          __typename: "ModelRelatedManyConnection",
          nextToken?: string | null,
        } | null,
        relatedOne?:  {
          __typename: "RelatedOne",
          id: string,
          content?: string | null,
          groups?: Array< string | null > | null,
          primaryId?: string | null,
          createdAt: string,
          updatedAt: string,
        } | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type CreateRelatedManyMutationVariables = {
  input: CreateRelatedManyInput,
  condition?: ModelRelatedManyConditionInput | null,
};

export type CreateRelatedManyMutation = {
  createRelatedMany?:  {
    __typename: "RelatedMany",
    id: string,
    content?: string | null,
    groups?: Array< string | null > | null,
    primaryId?: string | null,
    primary?:  {
      __typename: "Primary",
      id: string,
      content?: string | null,
      groups?: Array< string | null > | null,
      relatedMany?:  {
        __typename: "ModelRelatedManyConnection",
        items:  Array< {
          __typename: "RelatedMany",
          id: string,
          content?: string | null,
          groups?: Array< string | null > | null,
          primaryId?: string | null,
          createdAt: string,
          updatedAt: string,
        } | null >,
        nextToken?: string | null,
      } | null,
      relatedOne?:  {
        __typename: "RelatedOne",
        id: string,
        content?: string | null,
        groups?: Array< string | null > | null,
        primaryId?: string | null,
        primary?:  {
          __typename: "Primary",
          id: string,
          content?: string | null,
          groups?: Array< string | null > | null,
          createdAt: string,
          updatedAt: string,
        } | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type UpdateRelatedManyMutationVariables = {
  input: UpdateRelatedManyInput,
  condition?: ModelRelatedManyConditionInput | null,
};

export type UpdateRelatedManyMutation = {
  updateRelatedMany?:  {
    __typename: "RelatedMany",
    id: string,
    content?: string | null,
    groups?: Array< string | null > | null,
    primaryId?: string | null,
    primary?:  {
      __typename: "Primary",
      id: string,
      content?: string | null,
      groups?: Array< string | null > | null,
      relatedMany?:  {
        __typename: "ModelRelatedManyConnection",
        items:  Array< {
          __typename: "RelatedMany",
          id: string,
          content?: string | null,
          groups?: Array< string | null > | null,
          primaryId?: string | null,
          createdAt: string,
          updatedAt: string,
        } | null >,
        nextToken?: string | null,
      } | null,
      relatedOne?:  {
        __typename: "RelatedOne",
        id: string,
        content?: string | null,
        groups?: Array< string | null > | null,
        primaryId?: string | null,
        primary?:  {
          __typename: "Primary",
          id: string,
          content?: string | null,
          groups?: Array< string | null > | null,
          createdAt: string,
          updatedAt: string,
        } | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type DeleteRelatedManyMutationVariables = {
  input: DeleteRelatedManyInput,
  condition?: ModelRelatedManyConditionInput | null,
};

export type DeleteRelatedManyMutation = {
  deleteRelatedMany?:  {
    __typename: "RelatedMany",
    id: string,
    content?: string | null,
    groups?: Array< string | null > | null,
    primaryId?: string | null,
    primary?:  {
      __typename: "Primary",
      id: string,
      content?: string | null,
      groups?: Array< string | null > | null,
      relatedMany?:  {
        __typename: "ModelRelatedManyConnection",
        items:  Array< {
          __typename: "RelatedMany",
          id: string,
          content?: string | null,
          groups?: Array< string | null > | null,
          primaryId?: string | null,
          createdAt: string,
          updatedAt: string,
        } | null >,
        nextToken?: string | null,
      } | null,
      relatedOne?:  {
        __typename: "RelatedOne",
        id: string,
        content?: string | null,
        groups?: Array< string | null > | null,
        primaryId?: string | null,
        primary?:  {
          __typename: "Primary",
          id: string,
          content?: string | null,
          groups?: Array< string | null > | null,
          createdAt: string,
          updatedAt: string,
        } | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type CreateRelatedOneMutationVariables = {
  input: CreateRelatedOneInput,
  condition?: ModelRelatedOneConditionInput | null,
};

export type CreateRelatedOneMutation = {
  createRelatedOne?:  {
    __typename: "RelatedOne",
    id: string,
    content?: string | null,
    groups?: Array< string | null > | null,
    primaryId?: string | null,
    primary?:  {
      __typename: "Primary",
      id: string,
      content?: string | null,
      groups?: Array< string | null > | null,
      relatedMany?:  {
        __typename: "ModelRelatedManyConnection",
        items:  Array< {
          __typename: "RelatedMany",
          id: string,
          content?: string | null,
          groups?: Array< string | null > | null,
          primaryId?: string | null,
          createdAt: string,
          updatedAt: string,
        } | null >,
        nextToken?: string | null,
      } | null,
      relatedOne?:  {
        __typename: "RelatedOne",
        id: string,
        content?: string | null,
        groups?: Array< string | null > | null,
        primaryId?: string | null,
        primary?:  {
          __typename: "Primary",
          id: string,
          content?: string | null,
          groups?: Array< string | null > | null,
          createdAt: string,
          updatedAt: string,
        } | null,
        createdAt: string,
        updatedAt: string,
      } | null,
      createdAt: string,
      updatedAt: string,
    } | null,
    createdAt: string,
    updatedAt: string,
  } | null,
};

export type UpdateRelatedOneMutationVariables = {
  input: UpdateRelatedOneInput,
  condition?: ModelRelatedOneConditionInput | null,
};
// ...
export type DeleteRelatedOneMutationVariables = {
  input: DeleteRelatedOneInput,
  condition?: ModelRelatedOneConditionInput | null,
};